requirements.txt
numpy~=2.0.2
pillow
scikit-learn
